Italics and capitalization
Just a couple of updates to help bring this older post in line with the current style guidelines.

We don't usually italicize `bash` or `bash script`. I confirmed this by checking other posts in our blog. 

Recently, we agreed to capitalize Octopus terms only when referring to the feature, not a particular instance. For example, `Octopus Projects` vs. `Create a project`.