Suggested line edit
When I first read it, this phrase seemed incomplete. Perhaps it's my interpretation. Other readers may experience the same sticking point, so I wanted to flag it with you both.

Does the idea of not needing to follow along undermine the reasons to keep reading? You might be able to rephrase this in a way that doesn't put that question in the reader's mind.

I've suggested one possible approach to changing it. You could, instead, add one or more words between `need` and `to`, such as  `Octopus` or `your own Octopus instance`. Or ignore my suggestion and close the PR. 😅