Add link to `automate everything` in the handbook
I deliberated whether to mention the DevOps Engineer handbook in the link text, but our readers will quickly tire of "learn more in our handbook" if we overdo it. Here it might be cleaner to add the link to the matching keywords. I'm happy to follow any guidelines in this space – just let me know.